pyproject.toml
[tool.poetry]
name = "chat-service"
version = "0.1.0"
description = "chat-service"
authors = ["chestnut90 <cowzon90@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "chat_service"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
django-stubs = "^4.2.7"

[tool.black]
max-line-length = 120
force-exclude = 'migrations'

[tool.flake8]
max-line-length = 120
extend-ignore = """
    E203
    """
exclude = """
    .git,
    __pycache__,
    */migrations/*
    """

[tool.mypy]
strict = true
ignore_missing_imports = true

plugins = ["mypy_django_plugin.main" ]

[tool.django-stubs]
django_settings_module = "src.config.settings"
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
